api/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
asyncpg==0.29.0
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2
greenlet==3.0.3
tenacity==8.5.0
